build(test): fix tests config
